requirements.txt
bs4
django
djangorestframework
django-filter
pandas
psycopg2
python-dotenv
tqdm
